# Return elemwise logp in MvNormal.logp