The main stuff is in the first commit, the second is just whitespace.