@aseyboldt 2017-06-03
I also noticed while working on #2137, that the shape of `mu` can different when doing posterior sampling using `pm.sample_gp` and in the normal `pm.MvNormal.logp`. I'm a bit at a loss about why this might be.

```
with pm.Model() as model:
    l = pm.Uniform('l', 0, 100)
    
    log_s2_f = pm.Uniform('log_s2_f', lower=-10, upper=5)
    s2_f = pm.Deterministic('s2_f', tt.exp(log_s2_f))
    f_cov = s2_f * pm.gp.cov.ExpQuad(input_dim=1, lengthscales=l)
    
    log_s2_n = pm.Uniform('log_s2_n', lower=-10, upper=5)
    s2_n = pm.Deterministic('s2_n', tt.exp(log_s2_n))
    
    y_obs = pm.gp.GP('y_obs', cov_func=f_cov, sigma=s2_n, 
                     observed={'X': np.arange(10)[:, None], 
                               'Y': np.random.randn(10)[:, None]})

with model:
    trace = pm.sample()

with model:
    samples = pm.gp.sample_gp(trace, y_obs, X_values=np.arange(10)[:, None])
```

This now fails with 
```pytb
---------------------------------------------------------------------------
ValueError                                Traceback (most recent call last)
<ipython-input-5-557b9c433b8a> in <module>()
     17 
     18 with model:
---> 19     samples = pm.gp.sample_gp(trace, y_obs, X_values=np.arange(10)[:, None])

/home/adr/git/pymc3/pymc3/gp/gp.py in sample_gp(trace, gp, X_values, samples, obs_noise, model, random_seed, progressbar)
    141 
    142     try:
--> 143         samples = [gp_post.random(point=trace[idx]) for idx in indices]
    144     finally:
    145         if progressbar:

/home/adr/git/pymc3/pymc3/gp/gp.py in <listcomp>(.0)
    141 
    142     try:
--> 143         samples = [gp_post.random(point=trace[idx]) for idx in indices]
    144     finally:
    145         if progressbar:

/home/adr/git/pymc3/pymc3/distributions/multivariate.py in random(self, point, size)
    156             mu, cov = draw_values([self.mu, self.cov], point=point)
    157             if mu.shape != cov[0].shape:
--> 158                 raise ValueError("Shapes for mu an cov don't match")
    159 
    160             try:

ValueError: Shapes for mu an cov don't match
``