@aseyboldt 2017-06-03
#2111 could be related