@bwengals 2017-06-03
I guess that we should decide on the appropriate shape for the output of the mean functions, gp predictions, etc.  Do you think all the 1d vectors that show up in the math around this be coerced to have shape `(n,)` or `(n,1)`?