@aseyboldt 2017-06-03
`MvNormal` always interprets the last dimension as the one where the correlation happens. a (n,1) array would be interpreted as n independent samples from a 1d mvnormal.