@bwengals 2017-06-04
Oh ok!  So the mean functions should return shape `(n,)` instead of `(n,1)`.  I'll fix that as part of #2111 .  Could that info be in the docstring for `MvNormal`?