@twiecki 2017-06-09
LGTM, just need to resolve conflicts.