@junpenglao 2017-06-15
I think this is ready to merge now.