@aseyboldt 2017-06-15
Those timeouts...
It is possible that this PR slows some things down a bit. The previous code just computed a single logp for all observations, which can be a bit faster.