@junpenglao 2017-06-15
It is also because the travis of this build is still the old one (test_posterior was moved but not showed after)