@junpenglao 2017-06-20
@aseyboldt there are some significant slow down after this commit:
```python
with pm.Model():
    MvNormal('mu', mu=0, tau=np.eye(100), shape=(100,))
    pm.sample(10000)
```

Before
```
Auto-assigning NUTS sampler...
Initializing NUTS using ADVI...
Average Loss = 0.069045:   3%|▎         | 5159/200000 [00:00<00:18, 10305.70it/s]
Convergence archived at 5900
Interrupted at 5,900 [2%]: Average Loss = 2.1251
100%|██████████| 10500/10500 [00:09<00:00, 1061.38it/s]
```

After
```
Auto-assigning NUTS sampler...
Initializing NUTS using ADVI...
Average Loss = 0.065645:   3%|▎         | 5240/200000 [00:00<00:25, 7503.52it/s]
Convergence archived at 5900
Interrupted at 5,900 [2%]: Average Loss = 2.1251
100%|██████████| 10500/10500 [00:20<00:00, 515.30it/s]
```