@aseyboldt 2017-06-20
This is strange. I expected (and measured) a bit of slowdown if we look at many draws from a `MvNormal` at the same time, (so eg `shape=(1000, 100)`), but not for the case of a single value. The change that seems responsible is that we don't use `(delta_trans ** 2).sum()` anymore, but `(delta_trans ** ).sum(axis=0)`. I'd thought this wouldn't make much difference if `delta_trans` has shape `(100, 1)` but it does. From debugging this so far it looks like some optimization from theano is involved here:

Before:
```
<% time> <sum %> <apply time> <time per call> <type> <#call> <#apply> <Class name>
  53.9%    53.9%       4.105s       1.32e-05s     C   311856       3   theano.tensor.blas.Dot22
  32.3%    86.2%       2.459s       1.18e-05s     C   207904       2   theano.tensor.blas.Dot22Scalar
   4.5%    90.7%       0.342s       5.49e-07s     C   623712       6   theano.tensor.elemwise.Elemwise
   2.2%    92.9%       0.167s       4.02e-07s     C   415808       4   theano.tensor.basic.Reshape
   2.1%    94.9%       0.159s       3.81e-07s     C   415808       4   theano.tensor.elemwise.DimShuffle
   2.0%    97.0%       0.154s       1.48e-06s     C   103952       1   theano.compile.ops.DeepCopyOp
   0.8%    97.8%       0.064s       6.14e-07s     C   103952       1   theano.tensor.blas_c.CGemv
   0.8%    98.6%       0.063s       6.06e-07s     C   103952       1   theano.tensor.subtensor.Subtensor
   0.8%    99.4%       0.060s       5.79e-07s     C   103952       1   theano.tensor.elemwise.Sum
   0.6%   100.0%       0.045s       4.34e-07s     C   103952       1   theano.tensor.basic.AllocEmpty
```

After:
```
<% time> <sum %> <apply time> <time per call> <type> <#call> <#apply> <Class name>
  73.2%    73.2%      28.056s       9.98e-05s     C   281040       3   theano.tensor.blas_c.CGemv
  22.2%    95.4%       8.508s       3.03e-05s     C   281040       3   theano.tensor.blas.Dot22Scalar
   1.3%    96.7%       0.516s       5.51e-07s     C   936800      10   theano.tensor.elemwise.DimShuffle
   1.2%    97.9%       0.454s       8.08e-07s     C   562080       6   theano.tensor.elemwise.Elemwise
   1.1%    99.0%       0.404s       4.32e-06s     C    93680       1   theano.compile.ops.DeepCopyOp
   0.5%    99.5%       0.197s       7.00e-07s     C   281040       3   theano.tensor.basic.AllocEmpty
   0.3%    99.8%       0.110s       1.18e-06s     C    93680       1   theano.tensor.subtensor.Subtensor
   0.2%   100.0%       0.083s       8.81e-07s     C    93680       1   theano.tensor.elemwise.Sum
   ... (remaining 0 Classes account for   0.00%(0.00s) of the runtime)
```
From looking at the graph of the leapfrog step I'm not sure that theano is doing the right thing for MvNormal. I sometimes wish theano didn't try to be so smart. This makes tracking down what's happening really hard. I think we often end up with worse performance just because we have no idea what theano is doing.

What I'd like to do for the next release is to add the possibility for distributions to define a `logp_sum` method or so. The model already distinguishes between elemwise logp and sumed logp, so that wouldn't require changes outside the distributions, but allows us to use some more specialized code if we are only interested in the sum of the logp values. With this we can also overwrite the gradient of some logps with `OpFromGraph`. Maybe that can help us solve/avoid issues like this one. Unfortunately we can only use that once the next theano version is released, as that includes a necessary fix to `OpFromGraph`.

About the issue at hand: I'm not sure what to do about it. The previous behaviour led to bugs in loo and others, as they require elemwise logps. So just reverting doesn't seem like a good idea. And I don't know a simple fix for this that we could get into 3.1.
CC @twiecki @fonnesbeck