@twiecki 2017-06-20
I would say correctness trumps speed so we release with the slowdown (which only seems to happen for few data points anyhow) and work on your proposal for 3.2.